refactor(export): share query history serialization
